Catch at least *some* errors out of ajv.
Doesn't catch all of them, because even though AJV's strict mode is
documented to default to true, it also has a strictTypes setting which defaults
to 'log', meaning it will spew some messages to stderr (which currently bowtie
always marks as an uncaught error). Unclear yet precisely what to tweak.